Added ACSelect and LevelSelect to CharacterControlForm
