+++ src/components/characters/characterControl/CharacterControlDisplay.js
@@ -10,7 +10,7 @@ const CharacterControlDisplay = (props) => {
             {
                 props.form ? (
                     <div className="text-center">
-                        <div title="Armor Class">
+                        <div title="Armor Class" className="stat-with-icon">
                             <FaShieldAlt className='stats-icons' />
                             <div>{armorclass}</div>
                         </div>
@@ -20,16 +20,16 @@ const CharacterControlDisplay = (props) => {
                         </div>
                     </div>
                 ) : (
-                        <div className="character-control-display" onClick={props.onClick}>
-                            <div title="Armor Class" className="stat-with-icon">
-                                <FaShieldAlt className='stats-icons' />
-                                <div>{armorclass}</div>
-                            </div>
-                            <div title="Level" className="stat-with-icon" onClick={props.onClick}>
-                                <FaUserShield className='stats-icons' />
-                                <div>{level}</div>
-                            </div>
+                    <div className="character-control-display" onClick={props.onClick}>
+                        <div title="Armor Class" className="stat-with-icon">
+                            <FaShieldAlt className='stats-icons' />
+                            <div>{armorclass}</div>
                         </div>
+                        <div title="Level" className="stat-with-icon" onClick={props.onClick}>
+                            <FaUserShield className='stats-icons' />
+                            <div>{level}</div>
+                        </div>
+                    </div>
                     )
             }
         </React.Fragment>
