+++ src/components/characters/form/CharacterForm.js
@@ -1,0 +1,383 @@
+import React, { Component } from 'react'
+import {
+   MDBContainer,
+   MDBRow,
+   MDBCol,
+   MDBCard,
+   MDBCardBody,
+   MDBInput,
+   MDBBtn,
+   MDBIcon,
+   MDBAlert
+} from 'mdbreact'
+import { connect } from 'react-redux'
+import { createCharacter, patchCharacter } from '../../../redux/actions/characters'
+import validator from 'validator'
+import { validateAll } from 'indicative'
+import CharacterLevelSelect from './CharacterLevelSelect'
+import CharacterAcSelect from './CharacterAcSelect';
+
+// const levelOptions = (() => {
+//    const levels = Array.from(Array(21).keys())
+//    return levels.map(level => {
+//       if (level === 0)
+//          return (
+//             <option key={level} value='' disabled>
+//                {' '}
+//                Choose level...{' '}
+//             </option>
+//          )
+//       return (
+//          <option key={level} value={level}>
+//             {level}
+//          </option>
+//       )
+//    })
+// })()
+
+const armorClassOptions = (() => {
+   const armorClasses = Array.from(Array(31).keys())
+   return armorClasses.map(armorClass => {
+      if (armorClass === 0)
+         return (
+            <option key={armorClass} value='' disabled>
+               {' '}
+               Choose Armor Class...{' '}
+            </option>
+         )
+      return (
+         <option key={armorClass} value={armorClass}>
+            {armorClass}
+         </option>
+      )
+   })
+})()
+
+class CharacterForm extends Component {
+   constructor(props) {
+      super(props)
+      const updating = this.props.character ? this.props.character : false
+      const {
+         name,
+         level,
+         armorclass,
+         hitpoints,
+         maxhitpoints,
+         player
+      } = updating
+      this.state = {
+         updating,
+         name: updating ? name : '',
+         level: updating ? level : 1,
+         armorclass: updating ? armorclass : 10,
+         hitpoints: updating ? hitpoints : '',
+         maxhitpoints: updating ? maxhitpoints : '',
+         url: '',
+         characterPic: null,
+         player: updating ? player : false,
+         armorClassOptions,
+         style: {},
+         errors: {}
+      }
+   }
+
+   validateInput = () => {
+      //TODO make sure we validate hitpoints <= maxhitpoints (this is only relevant when state.updating)
+   }
+
+   handleKeyDown = event => {
+      switch (event.key) {
+         case 'Enter':
+            this.handleSubmit(this.props.toggleButtonNavigation)
+            break
+         case 'Escape':
+            this.handleCancel()
+            break
+         default:
+            break
+      }
+   }
+
+   handleChange = (type, value) => {
+      const newState = {}
+      newState[type] = value
+      this.setState({
+         ...newState
+      })
+   }
+
+   handleCreate = toggleCharacterNavigation => {
+      const data = this.state
+
+      const rules = {
+         name: 'required|string',
+         maxhitpoints: 'required|number'
+      }
+
+      const messages = {
+         required: 'Please fill in the {{ field }} field',
+         'maxhitpoints.number': 'Please type in a valid number'
+      }
+
+      validateAll(data, rules, messages)
+         .then(() => {
+            let picUrl = ''
+            if (!validator.isURL(this.state.url)) {
+               picUrl = null
+            } else {
+               picUrl = this.state.url
+            }
+            const payload = {
+               name: this.state.name,
+               level: this.state.level,
+               armorclass: this.state.armorclass,
+               player: this.state.player,
+               maxhitpoints: this.state.maxhitpoints,
+               user: this.props.User.userId,
+               characterPic: this.state.characterPic
+                  ? this.state.characterPic
+                  : picUrl
+            }
+
+            this.props.dispatch(
+               createCharacter(localStorage.getItem('DNDTOKEN'), payload)
+            )
+            toggleCharacterNavigation('Submit_Character')
+         })
+         .catch(errors => {
+            const formattedErrors = {}
+            errors.forEach(error => (formattedErrors[error.field] = error.message))
+            this.setState({
+               errors: formattedErrors
+            })
+         })
+   }
+
+   handleUpdate = () => {
+      const {
+         updating,
+         characterPic,
+         style,
+         levelOptions,
+         armorClassOptions,
+         url,
+         ...changedCharacter
+      } = this.state
+      const fieldsToUpdate = Object.entries(changedCharacter)
+         .filter(([key, value]) => {
+            return updating[key] !== value
+         })
+         .map(([propName, value]) => {
+            return {
+               propName,
+               value
+            }
+         })
+      if (fieldsToUpdate.length > 0) {
+         const maxHpIndex = fieldsToUpdate
+            .map(op => op.propName)
+            .indexOf('maxhitpoints')
+         if (maxHpIndex > -1 && !updating.player)
+            fieldsToUpdate.push({
+               propName: 'hitpoints',
+               value: fieldsToUpdate[maxHpIndex].value
+            })
+         this.props.dispatch(
+            patchCharacter(
+               localStorage.getItem('DNDTOKEN'),
+               fieldsToUpdate,
+               updating.request.url
+            )
+         )
+      } else {
+         this.handleCancel()
+      }
+   }
+
+   handleSubmit = toggleCharacterNavigation => {
+      if (!this.state.updating) {
+         this.handleCreate(toggleCharacterNavigation)
+      } else {
+         this.handleUpdate()
+      }
+      if (this.props.done) this.props.done()
+   }
+
+   handleCancel = () => {
+      if (this.props.done) this.props.done()
+   }
+
+   render() {
+      const {
+         name,
+         level,
+         armorclass,
+         hitpoints,
+         maxhitpoints,
+         player
+      } = this.state
+      const { toggleButtonNavigation } = this.props
+      return (
+         <MDBContainer style={this.state.style} className=''>
+            <MDBRow className='d-flex justify-content-center'>
+               <MDBCol md='8'>
+                  <MDBCard
+                     style={{ backgroundColor: 'transparent' }}
+                     className='create-character'
+                  >
+                     <MDBCardBody className='mx-4 d-row'>
+                        <div className='text-center'>
+                           <h3 className='mb-5'>
+                              <strong>
+                                 &nbsp;
+                      {this.state.updating ? (
+                                    `Edit '${this.state.updating.name}'`
+                                 ) : (
+                                       <MDBIcon
+                                          className='black-text'
+                                          icon='magic'
+                                          size='4x'
+                                       />
+                                    )}
+                              </strong>
+                           </h3>
+                        </div>
+                        <MDBInput
+                           label='Name'
+                           group
+                           containerClass='mb-0'
+                           required={true}
+                           onChange={e => 
+                              this.handleChange('name', e.target.value)
+                           }
+                           onKeyDown={e => 
+                              this.handleKeyDown(e)
+                           }
+                           value={name}
+                        />
+                        {this.state.errors.name && (
+                           <MDBAlert color='danger'>
+                              <MDBIcon icon='warning' />
+                              &nbsp;&nbsp;&nbsp;{this.state.errors.name}
+                           </MDBAlert>
+                        )}
+                        <CharacterLevelSelect
+                           value={level}
+                           onChange={value =>
+                              this.handleChange('level', value)
+                           }
+                        />
+                        <CharacterAcSelect
+                           value={armorclass}
+                           onChange={value => 
+                              this.handleChange('armorclass', value)
+                           }
+                        />
+                        {this.state.updating && this.state.updating.player && (
+                           <MDBInput
+                              className='browser-default custom-select'
+                              label='Hitpoints'
+                              containerClass='mb-0'
+                              onChange={e =>
+                                 this.handleChange('hitpoints', e.target.value)
+                              }
+                              onKeyDown={e => this.handleKeyDown(e)}
+                              value={hitpoints.toString()}
+                           />
+                        )}
+                        <MDBInput
+                           className='browser-default custom-select'
+                           label='Max Hitpoints'
+                           containerClass='mb-0'
+                           onChange={e =>
+                              this.handleChange('maxhitpoints', e.target.value)
+                           }
+                           onKeyDown={e => this.handleKeyDown(e)}
+                           value={maxhitpoints.toString()}
+                        />
+                        {this.state.errors.maxhitpoints && (
+                           <MDBAlert color='danger'>
+                              <MDBIcon icon='warning' />
+                              &nbsp;&nbsp;&nbsp;{this.state.errors.maxhitpoints}
+                           </MDBAlert>
+                        )}
+                        <MDBInput
+                           label='Player Character'
+                           className='mycheckbox'
+                           type='checkbox'
+                           id='checkbox'
+                           onChange={e => 
+                              this.handleChange('player', e.target.checked)
+                           }
+                           checked={player}
+                        />
+                        {!this.state.updating && (
+                           <div>
+                              <MDBInput
+                                 label='Photo URL'
+                                 className='text-center'
+                                 containerClass='mb-0'
+                                 onChange={e => 
+                                    this.handleChange('url', e.target.value)
+                                 }
+                                 onKeyDown={e => this.handleKeyDown(e)}
+                              />
+                              <MDBInput
+                                 type='file'
+                                 containerClass='mb-0'
+                                 onChange={e =>
+                                    this.handleChange('file', e.target.files[0])
+                                 }
+                                 onKeyDown={e => this.handleKeyDown(e)}
+                              />
+                           </div>
+                        )}
+                        <br />
+                        <div className='text-center'>
+                           <MDBBtn
+                              type='button'
+                              rounded
+                              color='black'
+                              className='btn-block z-depth-1a'
+                              onClick={() => 
+                                 this.handleSubmit(toggleButtonNavigation)
+                              }
+                           >
+                              {this.state.updating ? 'Save' : 'Create'}
+                           </MDBBtn>
+                        </div>
+                        <br />
+                        {this.state.updating && (
+                           <div className='text-center'>
+                              <MDBBtn
+                                 type='button'
+                                 rounded
+                                 color='black'
+                                 className='btn-block z-depth-1a'
+                                 onClick={() => this.handleCancel()}
+                              >
+                                 Cancel
+                              </MDBBtn>
+                           </div>
+                        )}
+                     </MDBCardBody>
+                  </MDBCard>
+               </MDBCol>
+            </MDBRow>
+         </MDBContainer>
+      )
+   }
+}
+
+function mapStateToProps({ User }) {
+   return {
+      User
+   }
+}
+
+// TODO: Move all these inline css to sepate file
+// const formHeaderStyle = {
+//   color: 'black'
+// }
+
+export default connect(mapStateToProps)(CharacterForm)
